requirements.txt
# aicli CLI — engine dependencies only
# No ChromaDB, no sentence-transformers, no SQLite

# LLM providers
anthropic>=0.40.0
openai>=1.50.0          # also used by DeepSeek and Grok (OpenAI-compat API)
google-genai>=1.0.0     # Gemini 2.0 (new SDK, replaces google-generativeai)

# CLI / UX
prompt_toolkit>=3.0.0
rich>=13.0.0

# Config / YAML
pyyaml>=6.0
pydantic>=2.0
python-dotenv>=1.0.0
